refactor: consolidate detection utilities to lib/shared (DRY)
- Create lib/shared/detection.ts as single source of truth for:
  - detectDeviceType, detectBrowser, detectOS
  - parseReferrer, parseReferrerWithLabel, parseUtmParams
  - isValidUrl, extractTitleFromUrl, normalizeUrl
  - REFERRER_LABELS constant

- Update lib/analytics.ts to re-export from shared module
- Update lib/url-utils.ts to re-export from shared module
- Update lib/analytics/utils.ts to use shared detection
- Update lib/analytics/constants.ts to re-export REFERRER_LABELS
- Remove ~150 lines of duplicated code across 3 files

All existing imports continue to work via re-exports.
